requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1
bcrypt==4.1.2
plotly==5.18.0
joblib==1.3.2
python-dotenv==1.0.1
openai>=1.12.0
python-jose==3.3.0
requests==2.31.0